Get event by id instead of traversing the list
Now that events are created with the id from the event in the email, we
can get the event directly instead of traversing the list of events.

This makes the script work with large calendars with more events than
the maximum that the list method returns. In that case, it may not find
the event, and try to create it, which would not work because the id
already exists.